requirements/base.txt


# Bleeding edge Django
django==1.10.3

# Configuration
django-environ==0.4.0



# Forms
django-braces==1.10.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.6

# Images
Pillow==3.4.2

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.28.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.7

# Redis support
django-redis==4.6.0
redis>=2.10.5





# Your custom requirements go here
https://github.com/ad-m/django-atom/archive/v0.11.0.tar.gz
